inference/vllm/pyproject.toml
[project]
name = "inference-vllm"
version = "0.1.0"
description = "Download and serve the Trinity-Mini model with vLLM + LoRA"
requires-python = ">=3.10, <3.13"
dependencies = [
    "modal",
    "vllm",
    "transformers",
    "requests",
    "python-dotenv",
]
